+++ strategy/__init__.py
@@ -1,5 +1,7 @@
 from strategy.ema_trend import SimpleEmaTrendStrategy
+from strategy.volatility_breakout import VolatilityBreakoutStrategy
 
 __all__ = [
     "SimpleEmaTrendStrategy",
+    "VolatilityBreakoutStrategy",
 ]